feat(EndpointEmpresa): Cria atribui as respostas com os códigos HTTP

+++ src/main/java/br/com/backendtestjava/backendtestjava/endpoint/EndpointEmpresa.java
@@ -3,6 +3,8 @@ package br.com.backendtestjava.backendtestjava.endpoint;
 import br.com.backendtestjava.backendtestjava.entity.Empresa;
 import br.com.backendtestjava.backendtestjava.respository.RepositoryEmpresa;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -26,23 +28,32 @@ public class EndpointEmpresa {
     }
 
     @PostMapping
-    public void addEmpresa(@RequestBody Empresa empresa) {
+    public ResponseEntity addEmpresa(@RequestBody Empresa empresa) {
         if (empresa.getId() != null) {
             empresa.setId(null);
         }
         repositoryEmpresa.save(empresa);
+        return new ResponseEntity(HttpStatus.CREATED);
     }
 
     @PutMapping(path = "/{id}")
-    public void editEmpresa(@PathVariable("id") Long id, @RequestBody Empresa empresa) {
+    public ResponseEntity editEmpresa(@PathVariable("id") Long id, @RequestBody Empresa empresa) {
         if (repositoryEmpresa.findById(id).isPresent()) {
             empresa.setId(id);
             repositoryEmpresa.save(empresa);
+            return new ResponseEntity(HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
     }
 
     @DeleteMapping(path = "/{id}")
-    public void deleteEmpresa(@PathVariable("id") Long id) {
-        repositoryEmpresa.deleteById(id);
+    public ResponseEntity deleteEmpresa(@PathVariable("id") Long id) {
+        if (repositoryEmpresa.findById(id).isPresent()) {
+            repositoryEmpresa.deleteById(id);
+            return new ResponseEntity(HttpStatus.OK);
+        } else {
+            return new ResponseEntity(HttpStatus.NOT_FOUND);
+        }
     }
 }
